requirements.txt
# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# AWS Bedrock Agent Runtime
aws-bedrock-agent-runtime>=1.0.0

# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.5.0
